+++ src/app/app.component.ts
@@ -1,5 +1,4 @@
 import { Component } from '@angular/core';
-import { Router } from '@angular/router';
 
 @Component({
     selector: 'app-root',
@@ -8,22 +7,4 @@ import { Router } from '@angular/router';
 })
 export class AppComponent {
     title = 'angularFile';
-
-    constructor(private readonly router: Router) {
-    }
-
-    public async downloadFile() {
-        const response = await fetch('assets/sample.jpg');
-        const blob =  await response.blob();
-        const url = window.URL.createObjectURL(blob);
-
-        const a = document.createElement('a');
-        document.body.appendChild(a);
-        a.setAttribute('style', 'display: none');
-        a.href = url;
-        a.download = 'Angular_sample_file';
-        a.click();
-        window.URL.revokeObjectURL(url);
-        this.router.navigate(['/homepage']);
-    }
 }
