+++ src/app/homepage/homepage.component.html
@@ -1,9 +1,9 @@
 <div class="container-fluid">
     <div class="container">
         <div class="row">
-
-
-            <div class="col-4"><p>homepage works!</p></div>
+            <div class="col-4">
+                <div (click)="downloadFile()">ダウンロード</div>
+            </div>
             <div class="col-4"><p>homepage works!</p></div>
             <div class="col-4"><p>homepage works!</p></div>
         </div>
