+++ src/app/homepage/homepage.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-homepage',
@@ -7,9 +8,24 @@ import { Component, OnInit } from '@angular/core';
 })
 export class HomepageComponent implements OnInit {
 
-  constructor() { }
+  constructor(private readonly router: Router) {
+  }
 
   ngOnInit(): void {
   }
 
+  public async downloadFile() {
+    const response = await fetch('assets/sample.jpg');
+    const blob =  await response.blob();
+    const url = window.URL.createObjectURL(blob);
+
+    const a = document.createElement('a');
+    document.body.appendChild(a);
+    a.setAttribute('style', 'display: none');
+    a.href = url;
+    a.download = 'Angular_sample_file';
+    a.click();
+    window.URL.revokeObjectURL(url);
+    this.router.navigate(['/homepage']);
+  }
 }
